@@ -22,7 +22,7 @@ public class AutoStartFix extends XposedModule {
             Class<?> BroadcastQueueInjector = XposedHelpers.findClass("com.android.server.am.BroadcastQueueInjector",loadPackageParam.classLoader);
             XposedUtils.findAndHookMethodAnyParam(BroadcastQueueInjector,"checkApplicationAutoStart",new XC_MethodHook() {
                 @Override
-                protected void beforeHookedMethod(MethodHookParam methodHookParam) throws Throwable {
+                protected void beforeHookedMethod(MethodHookParam methodHookParam) {
                     Intent intent = (Intent) XposedHelpers.getObjectField(methodHookParam.args[2], "intent");
                     if("com.google.android.c2dm.intent.RECEIVE".equals(intent.getAction())){
                         String target;
@@ -33,14 +33,6 @@ public class AutoStartFix extends XposedModule {
                         }
                         if(targetIsAllow(target)){
                             XposedHelpers.callStaticMethod(BroadcastQueueInjector,"checkAbnormalBroadcastInQueueLocked", methodHookParam.args[1], methodHookParam.args[0]);
-//                            XposedHelpers.callStaticMethod(
-//                                    XposedHelpers.findClass("android.miui.AppOpsUtils",loadPackageParam.classLoader),
-//                                    "noteApplicationAutoStart",
-//                                    XposedHelpers.getObjectField(methodHookParam.args[1], "mContext"),
-//                                    XposedUtils.getObjectFieldByPath(methodHookParam.args[3], "activityInfo.applicationInfo.packageName"),
-//                                    XposedUtils.getObjectFieldByPath(methodHookParam.args[3], "activityInfo.applicationInfo.uid"),
-//                                    "BroadcastQueueInjector#checkApplicationAutoStart"
-//                            );
                             methodHookParam.setResult(true);
                             printLog("Allow Auto Start: " + target);
                         }
@@ -52,11 +44,11 @@ public class AutoStartFix extends XposedModule {
         }
         try{
             // miui13
-            Class<?> BroadcastQueueImpl = XposedHelpers.findClass("com.android.server.am.BroadcastQueueImpl",loadPackageParam.classLoader);;
+            Class<?> BroadcastQueueImpl = XposedHelpers.findClass("com.android.server.am.BroadcastQueueImpl",loadPackageParam.classLoader);
             XposedUtils.findAndHookMethodAnyParam(BroadcastQueueImpl,"checkApplicationAutoStart",new XC_MethodHook() {
                 @Override
-                protected void beforeHookedMethod(MethodHookParam methodHookParam) throws Throwable {
-                    Intent intent = (Intent) XposedHelpers.getObjectField(methodHookParam.args[1], "intent");;
+                protected void beforeHookedMethod(MethodHookParam methodHookParam) {
+                    Intent intent = (Intent) XposedHelpers.getObjectField(methodHookParam.args[1], "intent");
                     if("com.google.android.c2dm.intent.RECEIVE".equals(intent.getAction())){
                         String target;
                         if (intent.getComponent() != null) {
@@ -84,7 +76,7 @@ public class AutoStartFix extends XposedModule {
             XposedUtils.findAndHookMethodAnyParam(AutoStartManagerService,"shouldInterceptService",new XC_MethodHook() {
 
                 @Override
-                protected void afterHookedMethod(MethodHookParam param) throws Throwable {
+                protected void afterHookedMethod(MethodHookParam param) {
                     Intent service = (Intent) param.args[0];
                     if("com.google.firebase.MESSAGING_EVENT".equals(service.getAction())){
                         printLog("Disable MIUI Intercept: " +
